Added note editing functionality. Changed the note color to be less bright. Added edit and delete buttons to notes. Tasks will now display their due date. Refactored the button bar component to use functions instead of state to fetch data. Changed to client side searching.
Known Issues:
The visual indicator for supabase updates breaks completely if the serach bar is used at any point. This is purely a visual bug and does not actually reflect whats going on behind the scenes.
The edit feature does not currently support changing the due date.
The delete button is not yet functional.